Started player class; made changes to other files.
I just realized that we don't need a values class since the player and
monster class can store these values. The monster class will now be
called the entity class because it will store all basic info of monsters
and the A.I. The main file now has  a simple menu for the demo.