Build 1.0 Released!
I decided to call all major updates to master into new build versions.
As such, the stat system works properly now and is set up for a player
menu that I will create in the future. I've also formatted the file
names to make it look neater, and edited the read me file.