Added Loot to Entity
Added a vector of loot to Entity and functions for it.

Changed the constructors for Dungeon so that it passes in two txt files.
The first txt file is the list of monsters and their stats/info. The
second txt file is the list of specific loot for that Dungeon. For now,
each entity in a dungeon will only have one loot.