Added Time System
Not sure if this is what you wanted. This Time class keeps track of the
day, the time, and the remaining time points it has. Once the time
points is empty, it will inform the player that it ran out of time and
will go on to the next day and reset time points back to 16 hours and
the time back to 8 AM.

You can test out my code in the time_test.cpp.  The idea is that you
call the update_time() function before an action.