Some Changes
JUST SHOWING SOMETHING.

Entity Changes:
1) So i noticed the change to a vector of Item pointers for loot instead
of Item objects, so I changed the add_loot function so that it pushes
back a pointer for the loot vector.
2) Since we dealing with pointers, I added in a destructor for Entity to
delete all the pointers in loot and make them equal to 0 (which is NULL)

Dungeon Changes:
1) Created an sample_entity file with 10 monsters. Note that for loot, i
only put in the Name of the loot and the Value just to keep things
simple for now. I can add in rarity and all that good stuff later.
2) Updated Dungeon so that you only pass in one txt file instead of two.
I made it so that it reads in loot if there is any.