Dungeon/Map
Dungeon Changes:
1) Added the map_file stirng into the constructor of
Dungeon. Was originally gonna read the map contents in fill_dungeon too,
but the map constructor is better and does it for me and I'd say it
makes it look
cleaner. The con is that you have to make two files for
one dungeon (dungeon info and the dungeon map), which could be tedious
and require
more files.
2) Created a map pointer in private for Dungeon.
That way, we can run dungeon in the dungeon class instead of map
class.
3) Because we now have a map pointer, I updated the clear_dun()
and destructor so that it deletes the pointer and points it to null to
avoid
any memory leaks/dangling pointers.
4) Like I said before, we're
running the dungeon in the actual dungeon class now. Created a
run_dungeon function, which is basically the same function as run in the
map class. SO DONT USE THE RUN FUNCTION IN THE MAP CLASS

Map
Changes:
1) Since I have a map pointer in Dungeon, I had to make get
functions to get the info about player position and goal position