Dungeon txt files
This is just the txt i used to test the functionality of my dungeon
functions.
1) I added spells and loot, even if some monsters dont have any i have
to add the :spells: / :loot: thing anyways because lazy.  I WILL ADD
MORE MONSTERS AND STUFF. And I will eventually change the txt file name
to something like "all_entity".
2) test_file.txt is the actual file that will be used to determine which
entities goes into whcih dungeon. Those files will have the Dungeon
name, Difficulty, and the names of the entities that are within that
dungeon.
2)