Changes
Entities: now have a vector of Consumables. So when it dies, it gives
the potions to the player.

Dungeon: Modified it so it reads in potions

Combat_System: Added stuff so that when the player wins, the entity
gives its potions to the player and it display correctly.
Too lazy to rename demo_dun_1.txt to all_entity.txt. I just created a
new file and deleted the old one. Still not all the monsters and still
not balanced but yeah.