ADDED SKILLS
Made the base class for skills. They only work against an entity when
dealt by the player, and is experimental still. Right now it can deal a
percentage of damage to an enemy or a range. It can heal, but I haven't
tested that out yet.

Combat and player classes have been modified for the addition of
classes.